@@ -4,28 +4,28 @@
   "private": true,
   "dependencies": {
     "axios": "^0.18.1",
-    "bootstrap": "^4.3.1",
-    "chart.js": "^2.8.0",
-    "firebase": "^6.5.0",
+    "bootstrap": "^4.4.1",
+    "chart.js": "^2.9.3",
+    "firebase": "^6.6.2",
     "font-awesome": "^4.7.0",
     "joi-browser": "^13.4.0",
     "jwt-decode": "^2.2.0",
-    "react": "^16.10.2",
-    "react-bootstrap": "^1.0.0-beta.9",
-    "react-chartjs-2": "^2.7.6",
-    "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.8.6",
+    "react": "^16.12.0",
+    "react-bootstrap": "^1.0.0-beta.16",
+    "react-chartjs-2": "^2.9.0",
+    "react-copy-to-clipboard": "^5.0.2",
+    "react-dom": "^16.12.0",
     "react-ga": "^2.7.0",
     "react-https-redirect": "^1.1.0",
     "react-images-upload": "^1.2.7",
-    "react-redux": "^7.1.0",
-    "react-router-dom": "^5.0.1",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.0.1",
-    "react-scroll-manager": "^1.0.0",
+    "react-scroll-manager": "^1.0.2",
     "react-sidebar": "^3.0.2",
-    "react-sizeme": "^2.6.7",
-    "react-toastify": "^5.3.1",
-    "redux": "^4.0.4",
+    "react-sizeme": "^2.6.12",
+    "react-toastify": "^5.5.0",
+    "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
     "semantic-ui-css": "^2.4.1",
